refactor: split inline-config into inline-config/ folder
