--- v-next/hardhat/src/internal/builtin-plugins/solidity-test/inline-config.ts
@@ -1,580 +1,0 @@
-import type {
-  BuildInfoAndOutput,
-  EdrArtifactWithMetadata,
-} from "./edr-artifacts.js";
-import type { SolidityBuildInfoOutput } from "../../../types/solidity/solidity-artifacts.js";
-import type {
-  ArtifactId,
-  TestFunctionConfigOverride,
-  TestFunctionOverride,
-} from "@nomicfoundation/edr";
-
-import {
-  HardhatError,
-  assertHardhatInvariant,
-} from "@nomicfoundation/hardhat-errors";
-import {
-  bytesToUtf8String,
-  bytesIncludesUtf8String,
-} from "@nomicfoundation/hardhat-utils/bytes";
-import { isObject } from "@nomicfoundation/hardhat-utils/lang";
-import {
-  kebabToCamelCase,
-  snakeToCamelCase,
-} from "@nomicfoundation/hardhat-utils/string";
-
-import { getFullyQualifiedName } from "../../../utils/contract-names.js";
-
-const HARDHAT_CONFIG_PREFIX = "hardhat-config:";
-const FORGE_CONFIG_PREFIX = "forge-config:";
-
-/** All supported inline config keys and their expected value types. */
-const KEY_TYPES: Record<string, "number" | "boolean"> = {
-  "fuzz.runs": "number",
-  "fuzz.maxTestRejects": "number",
-  "fuzz.showLogs": "boolean",
-  "fuzz.timeout": "number",
-  "invariant.runs": "number",
-  "invariant.depth": "number",
-  "invariant.failOnRevert": "boolean",
-  "invariant.callOverride": "boolean",
-  "invariant.timeout": "number",
-  allowInternalExpectRevert: "boolean",
-};
-
-/** Top-level key categories (e.g. "fuzz", "invariant", "allowInternalExpectRevert"). */
-const TOP_LEVEL_KEYS = [
-  ...new Set(Object.keys(KEY_TYPES).map((k) => k.split(".")[0])),
-];
-
-export interface RawInlineOverride {
-  inputSourceName: string;
-  contractName: string;
-  functionName: string;
-  functionSelector?: string; // from AST, hex without 0x prefix
-  key: string; // parsed camelCase key, without profile prefix
-  rawKey: string; // original key as written by the user, for error messages
-  rawValue: string;
-}
-
-/**
- * Extracts per-test inline configuration overrides from the NatSpec comments
- * in the solc AST.
- *
- * The same source file can appear in multiple build infos (as a transitive
- * dependency). A deduplication set prevents processing the same source twice.
- */
-export function getTestFunctionOverrides(
-  testSuiteArtifacts: EdrArtifactWithMetadata[],
-  buildInfosAndOutputs: BuildInfoAndOutput[],
-): TestFunctionOverride[] {
-  const allRawOverrides: RawInlineOverride[] = [];
-  const methodIdentifiersByContract = new Map<string, Record<string, string>>();
-  const processedSources = new Set<string>();
-
-  // Build lookup structures so we can skip irrelevant build infos
-  const testSuiteBuildInfoIds = new Set<string>();
-  const testSuiteSources = new Set<string>();
-  const artifactIdsByFqn = new Map<string, ArtifactId>();
-  for (const { edrArtifact, buildInfoId } of testSuiteArtifacts) {
-    testSuiteBuildInfoIds.add(buildInfoId);
-    testSuiteSources.add(edrArtifact.id.source);
-    const fqn = getFullyQualifiedName(
-      edrArtifact.id.source,
-      edrArtifact.id.name,
-    );
-    artifactIdsByFqn.set(fqn, edrArtifact.id);
-  }
-  const filteredBuildInfosAndOutputs = buildInfosAndOutputs.filter((bio) =>
-    testSuiteBuildInfoIds.has(bio.buildInfoId),
-  );
-
-  // Extract raw overrides and collect metadata for each source file.
-  // Only build infos referenced by testSuiteArtifacts are processed, and we
-  // only parse the output to get ASTs and method identifiers.
-  for (const buildInfoAndOutput of filteredBuildInfosAndOutputs) {
-    if (!buildInfoContainsInlineConfig(buildInfoAndOutput.buildInfo)) {
-      continue;
-    }
-
-    const buildInfoOutput: SolidityBuildInfoOutput = JSON.parse(
-      bytesToUtf8String(buildInfoAndOutput.output),
-    );
-
-    for (const inputSourceName of Object.keys(buildInfoOutput.output.sources)) {
-      if (!testSuiteSources.has(inputSourceName)) {
-        continue;
-      }
-      if (processedSources.has(inputSourceName)) {
-        continue;
-      }
-      processedSources.add(inputSourceName);
-
-      const source = buildInfoOutput.output.sources[inputSourceName];
-      const overrides = extractInlineConfigFromAst(source.ast, inputSourceName);
-      allRawOverrides.push(...overrides);
-
-      for (const [contractName, contractOutput] of Object.entries(
-        buildInfoOutput.output.contracts?.[inputSourceName] ?? {},
-      )) {
-        const fqn = getFullyQualifiedName(inputSourceName, contractName);
-        methodIdentifiersByContract.set(
-          fqn,
-          contractOutput.evm?.methodIdentifiers ?? {},
-        );
-      }
-    }
-  }
-
-  validateInlineOverrides(allRawOverrides);
-
-  // Group overrides by function. When the AST provides a functionSelector
-  // (public/external functions in solc >= 0.6.0), use it to distinguish
-  // overloaded functions. Otherwise fall back to function name only.
-  const overridesByFunction = new Map<string, RawInlineOverride[]>();
-  for (const override of allRawOverrides) {
-    const functionFqn = getFunctionFqn(
-      override.inputSourceName,
-      override.contractName,
-      override.functionName,
-    );
-    const groupKey =
-      override.functionSelector !== undefined
-        ? `${functionFqn}#${override.functionSelector}`
-        : functionFqn;
-    const existing = overridesByFunction.get(groupKey);
-    if (existing === undefined) {
-      overridesByFunction.set(groupKey, [override]);
-    } else {
-      existing.push(override);
-    }
-  }
-
-  // Build TestFunctionOverride objects
-  const testFunctionOverrides: TestFunctionOverride[] = [];
-  for (const [_groupKey, overrides] of overridesByFunction.entries()) {
-    const firstOverride = overrides[0];
-    const functionFqn = getFunctionFqn(
-      firstOverride.inputSourceName,
-      firstOverride.contractName,
-      firstOverride.functionName,
-    );
-    const contractFqn = `${firstOverride.inputSourceName}:${firstOverride.contractName}`;
-
-    const artifactId = artifactIdsByFqn.get(contractFqn);
-    assertHardhatInvariant(
-      artifactId !== undefined,
-      `Missing artifact id for "${contractFqn}"`,
-    );
-
-    // Use the AST-provided selector when available, otherwise fall back to
-    // resolving via methodIdentifiers.
-    let selector: string | undefined;
-    if (firstOverride.functionSelector !== undefined) {
-      selector = `0x${firstOverride.functionSelector}`;
-    } else {
-      const methodIdentifiers =
-        methodIdentifiersByContract.get(contractFqn);
-      assertHardhatInvariant(
-        methodIdentifiers !== undefined,
-        `Missing method identifiers for "${contractFqn}"`,
-      );
-
-      selector = resolveFunctionSelector(
-        methodIdentifiers,
-        firstOverride.functionName,
-      );
-    }
-
-    if (selector === undefined) {
-      throw new HardhatError(
-        HardhatError.ERRORS.CORE.SOLIDITY_TESTS
-          .INLINE_CONFIG_UNRESOLVED_SELECTOR,
-        { functionFqn },
-      );
-    }
-
-    testFunctionOverrides.push({
-      identifier: {
-        contractArtifact: artifactId,
-        functionSelector: selector,
-      },
-      config: buildConfigOverride(overrides),
-    });
-  }
-
-  return testFunctionOverrides;
-}
-
-/**
- * Returns true if the build info bytes contain either of the inline config
- * prefixes.
- */
-export function buildInfoContainsInlineConfig(
-  buildInfoBytes: Uint8Array,
-): boolean {
-  return (
-    bytesIncludesUtf8String(buildInfoBytes, HARDHAT_CONFIG_PREFIX) ||
-    bytesIncludesUtf8String(buildInfoBytes, FORGE_CONFIG_PREFIX)
-  );
-}
-
-/**
- * Extracts raw inline config overrides from a solc AST for a single source
- * file.
- */
-export function extractInlineConfigFromAst(
-  ast: unknown,
-  inputSourceName: string,
-): RawInlineOverride[] {
-  if (!isObject(ast) || ast.nodeType !== "SourceUnit") {
-    return [];
-  }
-
-  const results: RawInlineOverride[] = [];
-  const nodes: unknown[] = Array.isArray(ast.nodes) ? ast.nodes : [];
-  for (const node of nodes) {
-    if (!isObject(node) || node.nodeType !== "ContractDefinition") {
-      continue;
-    }
-
-    const contractName = node.name;
-    if (typeof contractName !== "string") {
-      continue;
-    }
-
-    const members: unknown[] = Array.isArray(node.nodes) ? node.nodes : [];
-    for (const member of members) {
-      if (!isObject(member) || member.nodeType !== "FunctionDefinition") {
-        continue;
-      }
-
-      const fnName = member.name;
-      if (
-        typeof fnName !== "string" ||
-        (!fnName.startsWith("test") && !fnName.startsWith("invariant"))
-      ) {
-        continue;
-      }
-
-      const fnSelector =
-        typeof member.functionSelector === "string"
-          ? member.functionSelector
-          : undefined;
-
-      const docText = extractDocText(member.documentation);
-      if (docText === undefined) {
-        continue;
-      }
-
-      for (const line of docText.split("\n")) {
-        const parsed = parseInlineConfigLine(
-          line,
-          inputSourceName,
-          contractName,
-          fnName,
-        );
-        if (parsed !== undefined) {
-          parsed.functionSelector = fnSelector;
-          results.push(parsed);
-        }
-      }
-    }
-  }
-
-  return results;
-}
-
-/**
- * Validates a list of raw inline overrides, checking for:
- * - Valid keys
- * - No duplicate keys for the same function
- * - Values of the expected type (numbers must be non-negative integers, booleans
- *   must be "true" or "false")
- *
- * Throws a HardhatError if any validation fails.
- */
-export function validateInlineOverrides(overrides: RawInlineOverride[]): void {
-  const seen = new Set<string>();
-
-  for (const {
-    inputSourceName,
-    contractName,
-    functionName,
-    functionSelector,
-    rawKey,
-    rawValue,
-    key,
-  } of overrides) {
-    const functionFqn = getFunctionFqn(
-      inputSourceName,
-      contractName,
-      functionName,
-    );
-
-    // Validate key
-    if (!Object.hasOwn(KEY_TYPES, key)) {
-      throw new HardhatError(
-        HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_INVALID_KEY,
-        {
-          key: rawKey,
-          validKeys: Object.keys(KEY_TYPES).join(", "),
-          functionFqn,
-        },
-      );
-    }
-
-    // Validate key matches test type
-    const dotIndex = key.indexOf(".");
-    if (dotIndex !== -1) {
-      const keyCategory = key.slice(0, dotIndex);
-      const isFuzzTest = functionName.startsWith("test");
-      const isInvariantTest = functionName.startsWith("invariant");
-
-      if (
-        (isFuzzTest && keyCategory === "invariant") ||
-        (isInvariantTest && keyCategory === "fuzz")
-      ) {
-        const testType = isFuzzTest ? "fuzz" : "invariant";
-        const validPrefix = isFuzzTest ? "fuzz." : "invariant.";
-        const validKeys = Object.keys(KEY_TYPES)
-          .filter((k) => k.startsWith(validPrefix) || !k.includes("."))
-          .join(", ");
-
-        throw new HardhatError(
-          HardhatError.ERRORS.CORE.SOLIDITY_TESTS
-            .INLINE_CONFIG_INVALID_KEY_FOR_TEST_TYPE,
-          {
-            key: rawKey,
-            functionFqn,
-            testType,
-            validKeys,
-          },
-        );
-      }
-    }
-
-    // Check for duplicates (include selector to allow same key on overloaded functions)
-    const functionId = functionSelector !== undefined
-      ? `${functionFqn}#${functionSelector}`
-      : functionFqn;
-    const dedupeKey = `${functionId}-${key}`;
-    if (seen.has(dedupeKey)) {
-      throw new HardhatError(
-        HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_DUPLICATE_KEY,
-        { key: rawKey, functionFqn },
-      );
-    }
-    seen.add(dedupeKey);
-
-    // Validate value type
-    const expectedType = KEY_TYPES[key];
-    if (expectedType === "number") {
-      if (
-        !/^(0|[1-9]\d*)$/.test(rawValue) ||
-        !Number.isSafeInteger(Number(rawValue))
-      ) {
-        throw new HardhatError(
-          HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_INVALID_VALUE,
-          {
-            value: rawValue,
-            key: rawKey,
-            expectedType: "non-negative integer",
-            functionFqn,
-          },
-        );
-      }
-    } else {
-      if (rawValue !== "true" && rawValue !== "false") {
-        throw new HardhatError(
-          HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_INVALID_VALUE,
-          {
-            value: rawValue,
-            key: rawKey,
-            expectedType: "boolean",
-            functionFqn,
-          },
-        );
-      }
-    }
-  }
-}
-
-/**
- * Finds the function selector for a given function name in the methodIdentifiers
- * map. Matches the first entry whose signature starts with `functionName(`.
- * Returns the selector prefixed with "0x".
- */
-export function resolveFunctionSelector(
-  methodIdentifiers: Record<string, string>,
-  functionName: string,
-): string | undefined {
-  const prefix = `${functionName}(`;
-
-  for (const [signature, selector] of Object.entries(methodIdentifiers)) {
-    if (signature.startsWith(prefix)) {
-      return `0x${selector}`;
-    }
-  }
-
-  return undefined;
-}
-
-/**
- * Converts a validated set of raw overrides into a TestFunctionConfigOverride
- * object.
- */
-export function buildConfigOverride(
-  overrides: RawInlineOverride[],
-): TestFunctionConfigOverride {
-  const config: Record<string, unknown> = {};
-  const fuzz: Record<string, unknown> = {};
-  const invariant: Record<string, unknown> = {};
-
-  for (const override of overrides) {
-    const expectedType = KEY_TYPES[override.key];
-    const parsed =
-      expectedType === "number"
-        ? Number(override.rawValue)
-        : override.rawValue === "true";
-
-    const dotIndex = override.key.indexOf(".");
-    if (dotIndex === -1) {
-      // Top-level key, like "allowInternalExpectRevert"
-      config[override.key] = parsed;
-    } else {
-      const section = override.key.slice(0, dotIndex);
-      const field = override.key.slice(dotIndex + 1);
-      const target = section === "fuzz" ? fuzz : invariant;
-      target[field] = field === "timeout" ? { time: parsed } : parsed;
-    }
-  }
-
-  if (Object.keys(fuzz).length > 0) {
-    config.fuzz = fuzz;
-  }
-  if (Object.keys(invariant).length > 0) {
-    config.invariant = invariant;
-  }
-
-  /* eslint-disable-next-line @typescript-eslint/consistent-type-assertions --
-  Cast is safe because we have validated all keys and value types in the previous step. */
-  return config as TestFunctionConfigOverride;
-}
-
-/**
- * Extracts the documentation text from a FunctionDefinition's documentation
- * field, which can be a StructuredDocumentation node, a plain string, or null.
- */
-export function extractDocText(doc: unknown): string | undefined {
-  if (
-    isObject(doc) &&
-    doc.nodeType === "StructuredDocumentation" &&
-    typeof doc.text === "string"
-  ) {
-    return doc.text;
-  } else if (typeof doc === "string") {
-    return doc;
-  } else {
-    return undefined;
-  }
-}
-
-/**
- * Parses a single line from a NatSpec comment and returns a RawInlineOverride
- * if the line contains a valid inline config directive. The line must start
- * with either "hardhat-config:" or "forge-config:", followed by a key=value pair.
- * Returns undefined if the line does not contain an inline config directive.
- */
-export function parseInlineConfigLine(
-  line: string,
-  inputSourceName: string,
-  contractName: string,
-  functionName: string,
-): RawInlineOverride | undefined {
-  // Strip leading whitespace and optional leading '*' from NatSpec text.
-  // Solc's StructuredDocumentation.text has delimiters (///, /**, */) removed.
-  // For /// comments, text has leading whitespace; for /** */ blocks, interior
-  // lines may start with " * ". The regex handles both styles.
-  const trimmedLine = line.replace(/^\s*\*?\s*/, "");
-  const functionFqn = getFunctionFqn(
-    inputSourceName,
-    contractName,
-    functionName,
-  );
-
-  let keyValueSegment: string;
-  let isForgeConfig: boolean;
-
-  if (trimmedLine.startsWith(HARDHAT_CONFIG_PREFIX)) {
-    keyValueSegment = trimmedLine.slice(HARDHAT_CONFIG_PREFIX.length).trim();
-    isForgeConfig = false;
-  } else if (trimmedLine.startsWith(FORGE_CONFIG_PREFIX)) {
-    keyValueSegment = trimmedLine.slice(FORGE_CONFIG_PREFIX.length).trim();
-    isForgeConfig = true;
-  } else {
-    return undefined;
-  }
-
-  const eqIndex = keyValueSegment.indexOf("=");
-  if (eqIndex === -1) {
-    throw new HardhatError(
-      HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_INVALID_SYNTAX,
-      {
-        line: trimmedLine,
-        functionFqn,
-      },
-    );
-  }
-
-  const rawKey = keyValueSegment.slice(0, eqIndex).trim();
-  let parsedKey = rawKey;
-  const rawValue = keyValueSegment.slice(eqIndex + 1).trim();
-
-  if (isForgeConfig) {
-    // Detect profile prefix: if the first dot-segment is NOT a known top-level
-    // category, treat it as a profile name.
-    const firstDot = rawKey.indexOf(".");
-    if (firstDot !== -1) {
-      const firstSegment = rawKey.slice(0, firstDot);
-      if (!TOP_LEVEL_KEYS.includes(firstSegment)) {
-        // It's a profile. Validate it.
-        const profile = firstSegment;
-        if (profile !== "default") {
-          throw new HardhatError(
-            HardhatError.ERRORS.CORE.SOLIDITY_TESTS.INLINE_CONFIG_UNSUPPORTED_PROFILE,
-            {
-              profile,
-              functionFqn,
-            },
-          );
-        }
-        // Strip the "default." prefix
-        parsedKey = rawKey.slice(firstDot + 1);
-      }
-    }
-  }
-
-  parsedKey = snakeToCamelCase(kebabToCamelCase(parsedKey));
-
-  return {
-    inputSourceName,
-    contractName,
-    functionName,
-    key: parsedKey,
-    rawKey,
-    rawValue,
-  };
-}
-
-/**
- * Constructs a fully qualified function name, in the format
- * "source.sol:ContractName#functionName".
- */
-export function getFunctionFqn(
-  inputSourceName: string,
-  contractName: string,
-  functionName: string,
-): string {
-  return `${getFullyQualifiedName(inputSourceName, contractName)}#${functionName}`;
-}
